+++ AddOns/PremadeGroupsFilter/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Premade Groups Filter
 
-## [6.9.2](https://github.com/0xbs/premade-groups-filter/tree/6.9.2) (2025-08-23)
-[Full Changelog](https://github.com/0xbs/premade-groups-filter/compare/6.9.1...6.9.2) [Previous Releases](https://github.com/0xbs/premade-groups-filter/releases)
+## [6.10.0](https://github.com/0xbs/premade-groups-filter/tree/6.10.0) (2025-10-09)
+[Full Changelog](https://github.com/0xbs/premade-groups-filter/compare/6.9.5...6.10.0) [Previous Releases](https://github.com/0xbs/premade-groups-filter/releases)
 
-- Fix lockout matching  
+- Add button to invert dungeon selection (fixes #344)  
+- Merge pull request #348 from Hollicsh/master  
+    Upd the .toc files  
+- Update PremadeGroupsFilter.toc  
+- Update PremadeGroupsFilter\_Mists.toc